% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aggregate_sparse.R
\name{aggregate_sparse}
\alias{aggregate_sparse}
\title{aggregate_sparse}
\usage{
aggregate_sparse(Expression_file, object = Clustering)
}
\arguments{
\item{Expression_file}{Gene Expression file as sparse matrix,}

\item{object}{A vector containing clustering or annotation}
}
\value{
Mean expression of genes in each group matrix
}
\description{
Aggregate a sparse matrix to compute the mean for each category of an object
}
