% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/C_normalisation.R
\name{C_normalisation}
\alias{C_normalisation}
\title{Normalize cell data based on cell size.}
\usage{
C_normalisation(y, scaling_factor)
}
\arguments{
\item{y}{A matrix of expression data.}

\item{scaling_factor}{A vector representing the area of each cell.}
}
\value{
A matrix of normalized cell data.
}
\description{
Normalize cell data based on cell size.
}
