% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calculate_log2fc.R
\name{calculate_log2fc}
\alias{calculate_log2fc}
\title{Calculate log2 fold change for a given gene across clusters.}
\usage{
calculate_log2fc(gene, Clustering)
}
\arguments{
\item{gene}{A string representing the gene of interest.}

\item{Clustering}{A vector indicating the cluster membership of each sample.}
}
\value{
A named vector of log2 fold changes for the specified gene.
}
\description{
Calculate log2 fold change for a given gene across clusters.
}
