% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Cell_clustering_function.R
\name{Cell_clustering_function}
\alias{Cell_clustering_function}
\title{Perform cell clustering using a K-nearest neighbors approach and Louvain algorithm.}
\usage{
Cell_clustering_function(Data_correction, K, metric_used, nThreads, resolution)
}
\arguments{
\item{Data_correction}{A matrix of reduced expression data.}

\item{K}{An integer specifying the number of neighbors to consider.}

\item{metric_used}{A string indicating the distance metric to use.}

\item{nThreads}{An integer specifying the number of threads for computation.}

\item{resolution}{A numeric value for the resolution parameter in clustering.}
}
\value{
A character vector indicating the cluster membership for each cell.
}
\description{
Perform cell clustering using a K-nearest neighbors approach and Louvain algorithm.
}
