% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/colvars_sparse.R
\name{colvars_sparse}
\alias{colvars_sparse}
\title{Compute variance of columns of a sparse matrix}
\usage{
colvars_sparse(mat)
}
\arguments{
\item{mat}{A sparse matrix}
}
\value{
A vector of values corresponding to the variance of each column of the input.
}
\description{
Compute variance of columns of a sparse matrix
}
