% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Compute_radius.R
\name{Compute_radius}
\alias{Compute_radius}
\title{Compute the radius of cells based on the specified method.}
\usage{
Compute_radius(Meta_data, Method)
}
\arguments{
\item{Meta_data}{A data frame containing cell metadata including area or volume.}

\item{Method}{A string specifying the method for radius calculation ('Xenium', 'CosMx', or 'Merfish').}
}
\value{
The updated Meta_data with calculated radius values.
}
\description{
Compute the radius of cells based on the specified method.
}
