% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Compute_variogram.R
\name{Compute_variogram}
\alias{Compute_variogram}
\title{Compute Variogram for Spatial Gene Expression Data}
\usage{
Compute_variogram(
  Expression_file,
  Meta_data,
  automatic_pad = TRUE,
  n_pad = 20,
  save_plot = FALSE,
  Output_path,
  ncores = 10
)
}
\arguments{
\item{Expression_file}{A data frame or matrix containing gene expression data.}

\item{Meta_data}{A data frame containing metadata for the cells, including their spatial coordinates.}

\item{automatic_pad}{A logical indicating whether to automatically determine the padding size. Default is TRUE.}

\item{n_pad}{An integer specifying the padding size. Default is 20. This is ignored if automatic_pad is TRUE.}

\item{save_plot}{A logical indicating whether to save the generated plots. Default is FALSE.}

\item{Output_path}{A character string specifying the path where output files (including plots) will be saved.}

\item{ncores}{An integer specifying the number of cores to use for parallel processing. Default is 10.}
}
\value{
A list containing:
   - Selected_genes: A vector of selected genes based on spatial variability.
   - Parameters: A data frame with parameters of the variogram model for each gene, including:
       - Sill
       - Range
       - Nugget
       - Alpha
       - Total variance
       - R-squared
       - C/C+n (ratio of explained variance) 
       - AIC
       - Model
       - Variogram values
}
\description{
This function computes the variogram for spatial gene expression data, allowing for
automatic padding, parallel processing, and optional saving of plots. It can handle
both data frames and summary expressions.
}
