% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Compute_variogram_multiple.R
\name{Compute_variogram_multiple}
\alias{Compute_variogram_multiple}
\title{Compute Variogram for Multiple Samples}
\usage{
Compute_variogram_multiple(
  Expression_file,
  Meta_data,
  automatic_pad = TRUE,
  n_pad = 20,
  save_plot = FALSE,
  Output_path,
  ncores = 10
)
}
\arguments{
\item{Expression_file}{A data frame or file path containing the expression data for the samples.}

\item{Meta_data}{A data frame containing metadata for the samples, including sample identifiers.}

\item{automatic_pad}{A logical value indicating whether to automatically pad the data. Default is TRUE.}

\item{n_pad}{An integer specifying the number of padding points to use if automatic padding is enabled. Default is 20.}

\item{save_plot}{A logical value indicating whether to save the generated plot. Default is FALSE.}

\item{Output_path}{A character string specifying the directory where output files (including plots) will be saved.}

\item{ncores}{An integer specifying the number of cores to use for parallel processing. Default is 10.}
}
\value{
A list containing the selected genes for each sample based on the computed variograms.
}
\description{
This function computes the variogram for multiple samples based on the provided expression data and metadata.
It generates an upset plot to visualize the shared genes across samples.
}
