% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Curate_data.R
\name{Curate_data}
\alias{Curate_data}
\title{Curate expression data and metadata based on quality control thresholds.}
\usage{
Curate_data(
  Expression_file,
  Meta_data,
  QC2_results,
  min_lib_size,
  max_lib_size,
  min_cell_radius,
  max_cell_radius
)
}
\arguments{
\item{Expression_file}{A matrix or dataframe of  gene expression data.}

\item{Meta_data}{A data frame containing metadata for each cell.}

\item{QC2_results}{A data frame containing quality control results.}

\item{min_lib_size}{Minimum total number of transcripts per cell threshold.}

\item{max_lib_size}{Maximum total number of transcripts per cell threshold.}

\item{min_cell_radius}{Minimum cell radius threshold.}

\item{max_cell_radius}{Maximum cell radius threshold.}
}
\value{
A list containing curated expression data and metadata.
}
\description{
Curate expression data and metadata based on quality control thresholds.
}
