% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Excess_variance_ratio_NB.R
\name{Excess_variance_ratio_NB}
\alias{Excess_variance_ratio_NB}
\title{Excess Variance Ratio computation using Negative Binomial Distribution}
\usage{
Excess_variance_ratio_NB(
  Expression_file,
  Output_path,
  Method,
  Tissue,
  P_value_threshold = 0.01,
  Ratio_threshold = 1.5
)
}
\arguments{
\item{Expression_file}{A data frame or matrix containing gene expression data, where rows represent genes
and columns represent samples.}

\item{Output_path}{A character string specifying the directory where the output PDF will be saved.}

\item{Method}{A character string specifying the method used for analysis.}

\item{Tissue}{A character string indicating the type of tissue being analyzed.}

\item{P_value_threshold}{A numeric value representing the threshold for statistical significance (default is 0.01).}

\item{Ratio_threshold}{A numeric value representing the threshold for the excess variance ratio (default is 1.5).}
}
\value{
A list containing:
     - Selected_genes: A character vector of gene names that show a significant excess variance ratio based on the specified thresholds.
     - Excess_variance_ratio: A numeric vector of excess variance ratios for each gene in the expression dataset.
}
\description{
This function analyzes gene expression data to identify genes exhibiting an excess variance ratio.
It employs a negative binomial model to calculate p-values and selects significant genes based on
user-defined thresholds for p-values and variance ratios and generates a plot of the results..
}
