% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Excess_zero_score_NB.R
\name{Excess_zero_score_NB}
\alias{Excess_zero_score_NB}
\title{Excess Zero Score computation using Negative Binomial Distribution}
\usage{
Excess_zero_score_NB(
  Expression_file,
  Output_path,
  Method,
  Tissue,
  P_value_threshold = 0.01,
  Delta_threshold = 0.01
)
}
\arguments{
\item{Expression_file}{A data frame or matrix containing gene expression data, where rows represent genes
and columns represent samples.}

\item{Output_path}{A character string specifying the directory where the output PDF will be saved.}

\item{Method}{A character string specifying the method used for analysis.}

\item{Tissue}{A character string indicating the type of tissue being analyzed.}

\item{P_value_threshold}{A numeric value representing the threshold for statistical significance (default is 0.01).}

\item{Delta_threshold}{A numeric value representing the threshold for the excess zero score (default is 0.01).}
}
\value{
A list containing:
   - Selected_genes: A data frame of genes that show a significant excess of zeros based on the specified thresholds.
   - Excess_zero_score: A numeric vector of excess zero scores for each gene in the expression dataset.
}
\description{
This function analyzes gene expression data to identify genes exhibiting an excess of zero counts.
It employs a negative binomial model to calculate p-values and selects significant genes based on
user-defined thresholds for p-values and delta scores and generates a plot of the results.
}
