% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fftshift.R
\name{fftshift}
\alias{fftshift}
\title{Shift the zero-frequency component to the center of the spectrum.}
\usage{
fftshift(input_matrix, dim = -1)
}
\arguments{
\item{input_matrix}{A matrix to be shifted.}

\item{dim}{An optional dimension to perform the shift (1 for rows, 2 for columns, -1 for both).}
}
\value{
The shifted matrix.
}
\description{
Shift the zero-frequency component to the center of the spectrum.
}
