% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Fraction_multiple_samples.R
\name{Fraction_multiple_samples}
\alias{Fraction_multiple_samples}
\title{This function takes the expression file and metadata and separates them into multiple ss based on the
metadata provided. It returns a list containing the expression data and metadata for each unique s.}
\usage{
Fraction_multiple_samples(Expression_file, Meta_data)
}
\arguments{
\item{Expression_file}{}

\item{Meta_data}{It must include a column named 's' that identifies the ss corresponding to the expression data.}
}
\value{
A list containing two elements:
\item{Expression_files}{A list of data frames, each corresponding to a unique s, filtered from the Expression_file.}
\item{Meta_data_files}{A list of data frames, each corresponding to the metadata for each unique s.}
}
\description{
This function takes the expression file and metadata and separates them into multiple ss based on the
metadata provided. It returns a list containing the expression data and metadata for each unique s.
}
