% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Geary_C_score.R
\name{Geary_C_score}
\alias{Geary_C_score}
\title{Compute Geary's C Score for Gene Expression Data}
\usage{
Geary_C_score(
  Expression_file,
  Meta_data,
  Output_path,
  Method,
  Tissue,
  pvalue_threshold = 0.01
)
}
\arguments{
\item{Expression_file}{A matrix or data frame where rows represent samples and columns represent genes.}

\item{Meta_data}{A data frame containing metadata with `cell_centroid_x` and `cell_centroid_y` representing the spatial coordinates of the cells.}

\item{Output_path}{A character string specifying the directory where the output PDF will be saved.}

\item{Method}{A character string specifying the method used for analysis.}

\item{Tissue}{A character string indicating the type of tissue being analyzed.}

\item{pvalue_threshold}{A numeric value indicating the threshold for adjusted p-values to select significant genes. Default is 0.01.}
}
\value{
A list containing:
   \item{Values}{A named numeric vector of Geary's C scores for each gene.}
   \item{Selected_genes}{A character vector of gene names that are significant based on the adjusted p-value threshold.}
   \item{P_values}{A numeric vector of adjusted p-values for each gene.}
}
\description{
This function calculates Geary's C score for each gene in the provided expression data,
using spatial information from the provided metadata. It also computes p-values and generates a plot of the results.
}
