% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Geary_C_score_multiple.R
\name{Geary_C_score_multiple}
\alias{Geary_C_score_multiple}
\title{Compute Geary's C for Multiple Samples
This function computes Geary's C score for multiple samples. It generates an upset plot to visualize the shared genes across samples.}
\usage{
Geary_C_score_multiple(
  Expression_file,
  Meta_data,
  Output_path,
  Method,
  Tissue,
  pvalue = 0.01
)
}
\arguments{
\item{Expression_file}{A list containing expression data for each sample.}

\item{Meta_data}{A list containing metadata for each sample.}

\item{Output_path}{A string specifying the directory where output files will be saved.}

\item{Method}{A string indicating the method to be used for computation.}

\item{Tissue}{A string specifying the tissue type for the analysis.}

\item{pvalue}{A numeric value indicating the p-value threshold for selecting genes. Default is 0.01.}
}
\value{
A list containing the selected genes for each sample.
}
\description{
Compute Geary's C for Multiple Samples
This function computes Geary's C score for multiple samples. It generates an upset plot to visualize the shared genes across samples.
}
