% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Get_isotropic_vario.R
\name{Get_isotropic_vario}
\alias{Get_isotropic_vario}
\title{Get the isotropic variogram from a given variogram matrix.}
\usage{
Get_isotropic_vario(Variogram)
}
\arguments{
\item{Variogram}{A matrix representing the variogram.}
}
\value{
A data frame with distances and corresponding variogram values.
}
\description{
Get the isotropic variogram from a given variogram matrix.
}
