% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Get_variogram_map.R
\name{Get_variogram_map}
\alias{Get_variogram_map}
\title{Get Variogram Map}
\usage{
Get_variogram_map(im)
}
\arguments{
\item{im}{A matrix representing the image for which the variogram map is to be computed.}
}
\value{
A matrix representing the variogram map of the input image.
}
\description{
This function computes the variogram map of a given image using the Fast Fourier Transform (FFT).
}
