% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ifftshift.R
\name{ifftshift}
\alias{ifftshift}
\title{Inverse FFT Shift}
\usage{
ifftshift(input_matrix, dim = -1)
}
\arguments{
\item{input_matrix}{A matrix to be shifted.}

\item{dim}{An integer indicating the dimension along which to perform the shift.
-1 for both dimensions, 1 for rows, and 2 for columns.}
}
\value{
A matrix after applying the inverse FFT shift.
}
\description{
This function performs an inverse FFT shift on a given matrix.
}
