% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Load_scimilarity_results.R
\name{Load_scimilarity_results}
\alias{Load_scimilarity_results}
\title{Load Scimilarity Results}
\usage{
Load_scimilarity_results(path, celltype_threshold = 0.01)
}
\arguments{
\item{path}{A string representing the file path to the directory containing the CSV file.}

\item{celltype_threshold}{A numeric value representing the threshold for cell type normalization.}
}
\value{
A vector of renamed cell types.
}
\description{
This function loads cell type predictions from a CSV file and normalizes the counts.
Cell types below a specified threshold are grouped into an "Other" category.
}
