% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Otsu_thresholding.R
\name{Otsu_thresholding}
\alias{Otsu_thresholding}
\title{Otsu Thresholding}
\usage{
Otsu_thresholding(x, number_bins = 100)
}
\arguments{
\item{x}{A numeric vector of data for which the threshold is to be determined.}

\item{number_bins}{An integer specifying the number of bins to use for histogram calculation.}
}
\value{
A numeric value representing the selected threshold.
}
\description{
This function implements in-house Otsu's method to determine an optimal threshold for a given dataset.
}
