% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pad_definitor.R
\name{pad_definitor}
\alias{pad_definitor}
\title{Pad Definitor}
\usage{
pad_definitor(Meta_data)
}
\arguments{
\item{Meta_data}{A data frame containing metadata with cell centroid coordinates.}
}
\value{
An integer representing the calculated padding size.
}
\description{
This function calculates an optimal padding size for rasterization based on the cell centroid coordinates in the metadata.
}
