% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Plot_FoV.R
\name{Plot_FoV}
\alias{Plot_FoV}
\title{Plot Field of View}
\usage{
Plot_FoV(FoV = 1)
}
\arguments{
\item{FoV}{An integer representing the field of view to be plotted.}
}
\value{
A plot of cell centroids for the specified FoV.
}
\description{
This function plots the cell centroids for a specified field of view (FoV).
}
