% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Plot_FoV_cluster.R
\name{Plot_FoV_cluster}
\alias{Plot_FoV_cluster}
\title{Plot Field of View with Clusters}
\usage{
Plot_FoV_cluster(object = Clustering)
}
\arguments{
\item{object}{Default is KNN clustering.}
}
\value{
A plot of cell centroids colored based on the classification.
}
\description{
This function plots the cell centroids colored by their classification.
}
