% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Plot_FoV_gene.R
\name{Plot_FoV_gene}
\alias{Plot_FoV_gene}
\title{Plot Field of View for a Gene}
\usage{
Plot_FoV_gene(gene)
}
\arguments{
\item{gene}{A string representing the gene for which the expression is to be plotted.}
}
\value{
A plot of cell centroids colored by gene expression.
}
\description{
This function plots the cell centroids colored by the expression of a specified gene.
}
