% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_variogram.R
\name{plot_variogram}
\alias{plot_variogram}
\title{Plot Variogram}
\usage{
plot_variogram(Sill, Range, Alpha, Gamma, Nugget, Model, filtered_x, R2)
}
\arguments{
\item{Sill}{Numeric value representing the sill of the variogram.}

\item{Range}{Numeric value representing the range of the variogram.}

\item{Alpha}{Numeric value used in the 'Finetuned exponential' model.}

\item{Nugget}{Numeric value representing the nugget effect.}

\item{Model}{Character string indicating the model type ('Constant', 'Exponential',
'Finetuned exponential', or 'Cosine exponential').}

\item{filtered_x}{A data frame containing the variogram data.}

\item{R2}{Numeric value representing the coefficient of determination.}
}
\value{
NULL This function does not return a value; it saves a PDF file.
}
\description{
This function plots the variogram based on the provided parameters and model type.
}
