% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/process_gene.R
\name{process_gene}
\alias{process_gene}
\title{Process to compute the Variogram of one Gene}
\usage{
process_gene(gene_name, sum_by_cut, n_pad, save_plot, Output_path)
}
\arguments{
\item{gene_name}{A string representing the name of the gene to be processed.}

\item{sum_by_cut}{A data frame containing the rasterized expression data.}

\item{n_pad}{A numeric value used to determine the size of unique elements in the rasterized data.}

\item{save_plot}{A logical value indicating whether to save the plot of the variogram.}

\item{Output_path}{A string specifying the path where output files should be saved.}
}
\value{
A numeric vector containing the following elements:
\item{Sill}{The sill of the best fitting model.}
\item{Range}{The range of the best fitting model (if applicable).}
\item{Nugget}{The nugget effect of the best fitting model (if applicable).}
\item{Alpha}{The alpha parameter of the best fitting model (if applicable).}
\item{Total_variance}{The total variance of the best fitting model.}
\item{R2}{The R-squared value of the best fitting model.}
\item{Prop}{The proportion of variance explained by the spatial structure of the best fitting model.}
\item{AIC}{The AIC value of the best fitting model.}
\item{Model}{A string indicating the name of the best fitting model.}
\item{Variogram}{A list of variogram values used in the fitting process.}
}
\description{
This function computes variogram for a gene and tries fitting different models to the variogram and 
selects the best model based on the Akaike Information Criterion (AIC).
}
