% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/QC_Gene_threshold.R
\name{QC_Gene_threshold}
\alias{QC_Gene_threshold}
\title{Quality Control Gene Thresholding}
\usage{
QC_Gene_threshold(Expression_file, Meta_data, Method, Tissue, Output_path)
}
\arguments{
\item{Expression_file}{A data frame containing gene expression data.}

\item{Meta_data}{A data frame containing metadata including radius information.}

\item{Method}{A character string specifying the method used for analysis.}

\item{Tissue}{A character string indicating the type of tissue being analyzed.}

\item{Output_path}{A character string specifying the directory where the output PDF will be saved.}
}
\value{
A list containing the Otsu threshold and the separability score.
}
\description{
This function performs quality control on gene expression data by applying Otsu's thresholding method
 and uses quality probes to calculate a separability score.
}
