% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/QC_RNA_size_threshold.R
\name{QC_RNA_size_threshold}
\alias{QC_RNA_size_threshold}
\title{QC_RNA_size_threshold}
\usage{
QC_RNA_size_threshold(Expression_file, Meta_data, Method, Tissue, Output_path)
}
\arguments{
\item{Expression_file}{A matrix or data frame containing gene expression data.}

\item{Meta_data}{A data frame containing metadata including a column named 'radius'.}

\item{Method}{A character string specifying the method used for analysis.}

\item{Tissue}{A character string indicating the type of tissue being analyzed.}

\item{Output_path}{A character string specifying the directory where the output PDF will be saved.}
}
\value{
NULL This function does not return a value; it saves a PDF file.
}
\description{
This function performs quality control on RNA expression data by analyzing the relationship between total RNA and cell size (radius).
It generates a 2D plot of RNA expression versus cell size and saves it as a PDF.
}
