% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Save_annotation_plot.R
\name{Save_annotation_plot}
\alias{Save_annotation_plot}
\title{Save Annotation Plot}
\usage{
Save_annotation_plot(Annotation_cells_renamed, Output_path, Method, Tissue)
}
\arguments{
\item{Annotation_cells_renamed}{A vector of annotation categories for cells. This should contain the annotations to visualize.}

\item{Output_path}{A character string specifying the directory path where the output PDF file will be saved.}

\item{Method}{A character string indicating the spatial method used.}

\item{Tissue}{A character string indicating the type of tissue.}
}
\value{
NULL This function does not return a value; it saves a PDF file.
}
\description{
This function generates a bar plot showing the frequency of different annotation categories
in the provided data and saves it as a PDF file.
}
