% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Save_boxplot.R
\name{Save_boxplot}
\alias{Save_boxplot}
\title{Save Boxplot of Gene Expression}
\usage{
Save_boxplot(
  Data_correction,
  object = Annotation_cells_renamed,
  gene = "MS4A1",
  Output_path,
  Method,
  Tissue
)
}
\arguments{
\item{Data_correction}{A data frame containing corrected gene expression data.}

\item{object}{A vector of cell type assignments.}

\item{gene}{A string representing the gene for which the boxplot is to be created.}

\item{Output_path}{A character string specifying the directory path where the output PDF plot will be saved.}

\item{Method}{A character string indicating the spatial method used.}

\item{Tissue}{A character string indicating the type of tissue.}
}
\value{
NULL This function does not return a value; it saves  a PDF file.
}
\description{
This function generates and saves a boxplot of gene expression across different cell types.
}
