% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Save_comparison_plots.R
\name{Save_comparison_plots}
\alias{Save_comparison_plots}
\title{Save Comparison Plots for Clustering and Annotation}
\usage{
Save_comparison_plots(
  Clustering,
  Annotation_cells_renamed,
  Umap_result,
  Output_path,
  Method,
  Tissue,
  scaling_factor = 2
)
}
\arguments{
\item{Clustering}{A vector indicating the clustering assignments for each cell.}

\item{Annotation_cells_renamed}{A vector indicating the annotations for each cell.}

\item{Umap_result}{A UMAP object containing UMAP coordinates for each cell.}

\item{Output_path}{A character string specifying the directory path where the output PDF plot will be saved.}

\item{Method}{A character string indicating the spatial method used.}

\item{Tissue}{A character string indicating the type of tissue.}

\item{scaling_factor}{A numeric value for scaling the point sizes in the plots. Default is 2.}
}
\value{
NULL This function does not return a value; it saves PNG files.
}
\description{
This function generates and saves comparison plots for clustering results and cell annotations.
It creates KNN and Scimilarity plots in both tissue atlas and UMAP formats.
}
