% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Save_dendogram.R
\name{Save_dendogram}
\alias{Save_dendogram}
\title{Save_dendogram}
\usage{
Save_dendogram(
  Clustering,
  Mean_expression,
  Annotation_cells_renamed,
  Output_path,
  Method,
  Tissue
)
}
\arguments{
\item{Clustering}{A vector or factor representing the clustering results.}

\item{Mean_expression}{A matrix or data frame containing mean expression values for each group.}

\item{Annotation_cells_renamed}{A vector of annotations for the cells.}

\item{Output_path}{A character string specifying the directory path where the output PDF plot will be saved.}

\item{Method}{A character string indicating the spatial method used.}

\item{Tissue}{A character string indicating the type of tissue.}
}
\value{
NULL This function does not return a value; it saves a PDF file.
}
\description{
This function generates and saves a dendrogram based on clustering results and Mean expression data.
It also creates a bar plot showing the frequency of annotations for each cluster and saves the 
combined output as a PDF.
}
