% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Save_geary_variance_plot.R
\name{Save_geary_variance_plot}
\alias{Save_geary_variance_plot}
\title{Save Geary Variance Plot}
\usage{
Save_geary_variance_plot(
  Variance_computation,
  Geary_computation,
  Output_path,
  Method,
  Tissue
)
}
\arguments{
\item{Variance_computation}{A list containing the results of variance computation.}

\item{Geary_computation}{A list containing the results of Geary computation.}

\item{Output_path}{A character string specifying the directory path where the output PDF plot will be saved.}

\item{Method}{A character string indicating the spatial method used.}

\item{Tissue}{A character string indicating the type of tissue.}
}
\value{
NULL This function does not return a value; it saves a PDF file.
}
\description{
This function generates a scatter plot comparing Variance scores and Geary scores for Selected genes.
The plot highlights Shared genes, Variance-only genes, and Geary-only genes with different colors.
}
