% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Save_heatmap_comparison.R
\name{Save_heatmap_comparison}
\alias{Save_heatmap_comparison}
\title{Save_heatmap_comparison}
\usage{
Save_heatmap_comparison(
  Clustering,
  Annotation_cells_renamed,
  Output_path,
  Method,
  Tissue
)
}
\arguments{
\item{Clustering}{A vector representing the clustering results.}

\item{Annotation_cells_renamed}{A vector of annotations for the cells.}

\item{Output_path}{A character string specifying the directory where the output PDF will be saved.}

\item{Method}{A character string indicating the spatial method used.}

\item{Tissue}{A character string indicating the type of tissue.}
}
\value{
NULL This function does not return a value; it saves a PDF file.
}
\description{
This function generates and saves a heatmap comparing clustering results with cell annotations.
It calculates the frequency of annotations for each cluster and visualizes the results in a heatmap.
}
