% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Save_heatmap_markers.R
\name{Save_heatmap_markers}
\alias{Save_heatmap_markers}
\title{Save Heatmap Markers}
\usage{
Save_heatmap_markers(
  Expression_file,
  object = Clustering,
  Output_path,
  Method,
  Tissue,
  name_object = "Clustering"
)
}
\arguments{
\item{Expression_file}{A data frame or a sparse matrix containing gene expression data with genes as columns and cells as rows.}

\item{object}{A clustering object that defines the clusters for the samples. This is typically a vector 
or factor indicating the cluster assignment for each sample.}

\item{Output_path}{A character string specifying the directory where the output PDF file will be saved.}

\item{Method}{A character string indicating the spatial method used.}

\item{Tissue}{A character string indicating the type of tissue.}
}
\value{
NULL This function does not return a value; it saves a PDF file.
}
\description{
This function generates a heatmap of the top marker genes for each cluster based on the provided expression data.
It calculates the log2 fold change for each gene and selects the top 5 genes with the highest log2 fold change
for each cluster. The resulting heatmap is saved as a PDF file.
}
