% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Save_umap.R
\name{Save_umap}
\alias{Save_umap}
\title{Save UMAP Plot}
\usage{
Save_umap(
  Umap_result,
  object = Clustering,
  Output_path,
  Method,
  Tissue,
  name_object = "Clustering"
)
}
\arguments{
\item{Umap_result}{A matrix containing UMAP results.}

\item{object}{A vector of clustering results corresponding to the UMAP points.}

\item{Output_path}{Character string specifying the output directory where to save the plot.}

\item{Method}{A character string indicating the spatial method used.}

\item{Tissue}{A character string indicating the type of tissue.}
}
\value{
NULL This function does not return a value; it saves a PNG file.
}
\description{
This function saves a UMAP plot of colored points, based on clustering results, to a PNG file.
}
