% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Select_genes.R
\name{Select_genes}
\alias{Select_genes}
\title{Select and visualize shared genes}
\usage{
Select_genes(
  Selected_objects = list(Variance_genes, Zero_genes),
  Selected_names = c("Variance_genes", "Zero_genes")
)
}
\arguments{
\item{Selected_objects}{A list containing vectors of gene names. Each vector represents a set of genes (e.g., variance genes, zero genes).}

\item{Selected_names}{A character vector containing names for the gene sets corresponding to the vectors in `Selected_objects`.}
}
\value{
A vector of unique shared genes across the provided gene sets.
}
\description{
This function takes a list of gene sets and generates an UpSet plot to visualize the shared genes among them.
}
