% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/string.to.colors.R
\name{string.to.colors}
\alias{string.to.colors}
\title{Convert Strings to Colors}
\usage{
string.to.colors(string, colors = NULL)
}
\arguments{
\item{string}{A character vector or factor containing the strings to be converted
to colors.}

\item{colors}{A character vector of colors to be used for the unique elements
in `string`. If NULL, a rainbow palette will be generated.}
}
\value{
A character vector of colors corresponding to the input strings.
}
\description{
This function maps unique strings to colors. In absence of a provided color palette,
it will generate a rainbow palette.
}
\examples{
annotation = c('B cells', 'T cells' , 'B cells', 'pDCs')
colors = string.to.colors(annotation)
print(colors)
[1] "#FF0000" "#00FF00" "#FF0000" "#0000FF"

}
