% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Umap_maker.R
\name{Umap_maker}
\alias{Umap_maker}
\title{UMAP Maker}
\usage{
Umap_maker(PCA_data, n_neighbors = 30, metric = "correlation")
}
\arguments{
\item{PCA_data}{A data frame containing PCA results with a column 'u' containing PCA coordinates.}

\item{n_neighbors}{Integer specifying the number of neighbors to consider for UMAP.}

\item{metric}{Character string indicating the distance metric to use (default is "correlation").}
}
\value{
A matrix of UMAP results.
}
\description{
This function generates UMAP embeddings from PCA data.
}
