% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Save_tissue_visualization.R
\name{Save_tissue_visualization}
\alias{Save_tissue_visualization}
\title{Save Tissue Visualization}
\usage{
Save_tissue_visualization(
  Meta_data,
  object = Clustering,
  Output_path,
  Method,
  Tissue,
  name_object = "Clustering",
  scaling_factor = 1.5
)
}
\arguments{
\item{Meta_data}{A data frame containing the metadata for the cells with
the x and y coordinates (columns named `cell_centroid_x` and `cell_centroid_y`).}

\item{object}{A vector representing classification of cells. Default is KNN Clustering result.}

\item{Output_path}{A character string specifying the directory where the output PNG file will be saved.}

\item{Method}{A character string indicating the spatial method used.}

\item{Tissue}{A character string indicating the type of tissue.}

\item{scaling_factor}{A numeric value that determines the size of the points in the plot. Default is 1.5.}
}
\value{
NULL This function does not return a value; it saves a PNG file.
}
\description{
This function generates a visualization of tissue based on cell centroids and clusters.
It saves the plot as a PNG file in the specified output path.
}
